+++ core/src/main/kotlin/io/github/serpro69/kfaker/dictionary/Dictionary.kt
@@ -28,6 +28,7 @@ internal enum class CategoryName {
     BARCODE,
     BASKETBALL,
     BEER,
+    BIG_BANG_THEORY,
     BLOOD,
     BOJACK_HORSEMAN,
     BOOK,